chore: move ParseDelimiter
