+++ pkg/shares/utils.go
@@ -93,3 +93,34 @@ func zeroPadIfNecessary(share []byte, width int) (padded []byte, bytesOfPadding 
 	share = append(share, padding...)
 	return share, missingBytes
 }
+
+// ParseDelimiter finds and returns the length delimiter of the share provided
+// while also removing the delimiter bytes from the input. ParseDelimiter
+// applies to both compact and sparse shares. Input should not contain the
+// namespace ID or info byte of a share.
+func ParseDelimiter(input []byte) (inputWithoutLengthDelimiter []byte, dataLength uint64, err error) {
+	if len(input) == 0 {
+		return input, 0, nil
+	}
+
+	l := binary.MaxVarintLen64
+	if len(input) < binary.MaxVarintLen64 {
+		l = len(input)
+	}
+
+	delimiter, _ := zeroPadIfNecessary(input[:l], binary.MaxVarintLen64)
+
+	// read the length of the data
+	r := bytes.NewBuffer(delimiter)
+	dataLen, err := binary.ReadUvarint(r)
+	if err != nil {
+		return nil, 0, err
+	}
+
+	// calculate the number of bytes used by the delimiter
+	lenBuf := make([]byte, binary.MaxVarintLen64)
+	n := binary.PutUvarint(lenBuf, dataLen)
+
+	// return the input without the length delimiter
+	return input[n:], dataLen, nil
+}
